+++ NEWS.md
@@ -60,6 +60,8 @@ New library features
 * `sort(keys(::Dict))` and `sort(values(::Dict))` now automatically collect, they previously threw ([#56978]).
 * `Base.AbstractOneTo` is added as a supertype of one-based axes, with `Base.OneTo` as its subtype ([#56902]).
 * `takestring!(::IOBuffer)` removes the content from the buffer, returning the content as a `String`.
+* New forms `@allocated inside f(...)` and `@allocations inside f(...)` to measure allocations inside
+  a function, excluding arguments and calling context ([#59278]).
 
 Standard library changes
 ------------------------